requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.0
kiwisolver==1.4.9
matplotlib==3.10.6
numpy==2.3.3
scipy>=1.16.2
packaging==25.0
pillow==11.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
six==1.17.0
